// devtools-ui/src/api/model/getModuleDetailsResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Awilixify Devtools API
 * API for inspecting module graphs, traces, and provider impact
 * OpenAPI spec version: 1.0.0
 */
import type { ModuleGraphNode } from "./moduleGraphNode";
import type { ModuleGraphRoute } from "./moduleGraphRoute";

export interface GetModuleDetailsResponse {
	availableCommandPreHandlers: string[];
	availableInitializers: string[];
	availableInterceptors: string[];
	availableQueryPreHandlers: string[];
	globalModules: string[];
	importedModules: string[];
	module: ModuleGraphNode;
	routes: ModuleGraphRoute[];
	usedByModules: string[];
}