// devtools-ui/src/api/model/moduleGraphNodeBase.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Awilixify Devtools API
 * API for inspecting module graphs, traces, and provider impact
 * OpenAPI spec version: 1.0.0
 */
import type { DynamicModuleInfo } from "./dynamicModuleInfo";
import type { ModuleGraphNodeBaseLifetimeTypes } from "./moduleGraphNodeBaseLifetimeTypes";
import type { ModuleGraphNodeBaseProviderAllowCircular } from "./moduleGraphNodeBaseProviderAllowCircular";
import type { ModuleGraphNodeBaseProviderDependencies } from "./moduleGraphNodeBaseProviderDependencies";
import type { ModuleGraphNodeBaseProviderEager } from "./moduleGraphNodeBaseProviderEager";
import type { ModuleGraphNodeBaseProviderInitAfter } from "./moduleGraphNodeBaseProviderInitAfter";
import type { ModuleGraphNodeKind } from "./moduleGraphNodeKind";

export interface ModuleGraphNodeBase {
	id: string;
	name: string;
	baseName?: string;
	dynamic?: DynamicModuleInfo;
	grouped: boolean;
	familyInstanceCount: number;
	instanceCount: number;
	kind: ModuleGraphNodeKind;
	dependencyCount: number;
	dependentCount: number;
	providers: string[];
	providerAllowCircular: ModuleGraphNodeBaseProviderAllowCircular;
	providerDependencies: ModuleGraphNodeBaseProviderDependencies;
	providerEager: ModuleGraphNodeBaseProviderEager;
	providerInitAfter: ModuleGraphNodeBaseProviderInitAfter;
	lifetimeTypes: ModuleGraphNodeBaseLifetimeTypes;
	exports: string[];
	controllers: string[];
	queryHandlers: string[];
	commandHandlers: string[];
	queryPreHandlers: string[];
	queryPreHandlerExports: string[];
	commandPreHandlers: string[];
	commandPreHandlerExports: string[];
	interceptors: string[];
	interceptorExports: string[];
	initializers: string[];
	initializerExports: string[];
}